Documentation cleanup, added time-window sampling to TimeSeries

+++ src/main/java/net/johnewart/chronos/Frequency.java
@@ -3,15 +3,15 @@ package net.johnewart.chronos;
 import java.util.concurrent.TimeUnit;
 
 public class Frequency {
-    public final long offset;
+    public final int offset;
     public final TimeUnit timeUnit;
 
-    public Frequency(long offset, TimeUnit timeUnit) {
+    public Frequency(int offset, TimeUnit timeUnit) {
         this.offset = offset;
         this.timeUnit = timeUnit;
     }
 
-    public static Frequency of(long offset, TimeUnit timeUnit) {
+    public static Frequency of(int offset, TimeUnit timeUnit) {
         return new Frequency(offset, timeUnit);
     }
